+++ ton_client/src/json_interface/modules.rs
@@ -113,6 +113,10 @@ fn register_crypto(handlers: &mut RuntimeHandlers) {
         crate::crypto::nacl_sign_detached,
         crate::crypto::nacl::nacl_sign_detached_api,
     );
+    module.register_sync_fn(
+        crate::crypto::nacl_sign_detached_verify,
+        crate::crypto::nacl::nacl_sign_detached_verify_api,
+    );
 
     module.register_sync_fn_without_args(
         crate::crypto::nacl_box_keypair,
@@ -301,7 +305,10 @@ fn register_boc(handlers: &mut RuntimeHandlers) {
         crate::boc::blockchain_config::get_blockchain_config_api,
     );
     module.register_sync_fn(crate::boc::get_boc_hash, crate::boc::hash::get_boc_hash_api);
-    module.register_sync_fn(crate::boc::get_code_from_tvc, crate::boc::tvc::get_code_from_tvc_api);
+    module.register_sync_fn(
+        crate::boc::get_code_from_tvc,
+        crate::boc::tvc::get_code_from_tvc_api,
+    );
     module.register();
 }
 
@@ -317,10 +324,7 @@ fn register_net(handlers: &mut RuntimeHandlers) {
     module.register_type::<crate::net::OrderBy>();
     module.register_type::<crate::net::SortDirection>();
 
-    module.register_async_fn(
-        crate::net::query,
-        crate::net::queries::query_api,
-    );
+    module.register_async_fn(crate::net::query, crate::net::queries::query_api);
     module.register_async_fn(
         crate::net::query_collection,
         crate::net::queries::query_collection_api,
@@ -329,7 +333,10 @@ fn register_net(handlers: &mut RuntimeHandlers) {
         crate::net::wait_for_collection,
         crate::net::queries::wait_for_collection_api,
     );
-    module.register_async_fn(crate::net::unsubscribe, crate::net::subscriptions::unsubscribe_api);
+    module.register_async_fn(
+        crate::net::unsubscribe,
+        crate::net::subscriptions::unsubscribe_api,
+    );
     module.register_async_fn_with_callback(
         super::net::subscribe_collection,
         super::net::subscribe_collection_api,